add MapMarkers component from Map
